requirements.txt
gtts 2.4.0
ffmpeg-2023-12-18-git-be8a4f80b9-full_build
pydub 0.25.1
playsound-1.3.0
TikTok-Voice-TTS
moviepy 1.0.3
ImageMagick 7.1.1-25-Q16-HDRI